move promisify tests to test/util.ts
